+++ src/UiAdminBase/Action/Route/RouteAdd/RouteAddPayloadCollection.php
@@ -1,0 +1,29 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Ui\Admin\Base\Action\Route\RouteAdd;
+
+use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
+use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
+use Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection;
+use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
+
+/**
+ * @extends AbstractObjectCollection<RouteAddPayload>
+ */
+final class RouteAddPayloadCollection extends AbstractObjectCollection implements AttachmentAwareInterface
+{
+    use AttachmentAwareTrait;
+
+    public function __construct(iterable $items = [])
+    {
+        parent::__construct($items);
+        $this->attachments = new AttachmentCollection();
+    }
+
+    protected function getT(): string
+    {
+        return RouteAddPayload::class;
+    }
+}